feat(RESOURCE): Game page Keywords and locators
There was no Keyword that waited for the page to become interactable after the browser loaded the initial document.
There was no Keyword that could close the cookie policy dialog and the preroll ad.
There were no supporting locators for the cookie dialog, preroll container, google ads iframe. and the game container.