feat(TEST): Solutions for puzzle
There was no structured format that provided an example of how solutions should specified for puzzles.
Although this approach, in which puzzle solutions are explicitly specified as test data, may require more manual input than other alternatives, it ensures that the clues and solutions loaded in a puzzle are always the ones that have been previously setup for the target day.